Modularize processing functions, add generic helpers, and improve test coverage
Extracts domain-specific functions from `spinlab.processing.helpers` into new, dedicated modules, improving code organization and maintainability. Repurposes `helpers.py` for common utilities and backward compatibility exports.

Introduces new unit tests for all refactored modules and includes minor fixes for normalization (region handling) and SNR calculations (division by zero).
