+++ spinlab/processing/__init__.py
@@ -2,11 +2,18 @@
 
 from .align import *
 from .apodization import *
+from .axis import *
+from .complex_data import *
+from .enhancement import *
 from .fft import *
-from .helpers import *
 from .integration import *
+from .modulation import *
+from .normalization import *
 from .offset import *
 from .phase import *
 from .conversion import *
 from .average import *
 from .interpolation import *
+from .smoothing import *
+from .snr import *
+from .helpers import *
